Create luxmus.c
Patch or Rewrite of mus.c - Ubuntu adaptation of the chromebook version of source code.. please note mus.c will not work outside of the cromebook simulation in part due to, cross-platform library issues. Maybe this is googles hack job of defanging the beast we all love GCC. Chrome OS, the operating system on Chromebooks, is based on Linux but has certain restrictions that limit access to system-level functionalities. This can make it challenging to install and use development tools like GCC, especially if you're looking to perform tasks that require low-level system access.

Chrome OS emphasizes simplicity and security, which means users may need to navigate additional layers of abstraction or use tools like Linux containers to access the full suite of development tools available on Linux systems. Is it security or something other  then just don't be evil ? While it's certainly possible to use GCC on a Chromebook with some effort and potentially additional tools like Crostini (Linux container support on Chrome OS), the experience may not be as um, leak free or seamless as on a traditional Linux development environment.

After all free as-in freedom!